+++ beacon-chain/state/stateutil/field_root_test.go
@@ -1,4 +1,4 @@
-package v3
+package stateutil
 
 import (
 	"testing"
@@ -7,17 +7,17 @@ import (
 )
 
 func TestArraysTreeRoot_OnlyPowerOf2(t *testing.T) {
-	_, err := nocachedHasher.arraysRoot([][]byte{}, 1, "testing")
+	_, err := NocachedHasher.arraysRoot([][]byte{}, 1, "testing")
 	assert.NoError(t, err)
-	_, err = nocachedHasher.arraysRoot([][]byte{}, 4, "testing")
+	_, err = NocachedHasher.arraysRoot([][]byte{}, 4, "testing")
 	assert.NoError(t, err)
-	_, err = nocachedHasher.arraysRoot([][]byte{}, 8, "testing")
+	_, err = NocachedHasher.arraysRoot([][]byte{}, 8, "testing")
 	assert.NoError(t, err)
-	_, err = nocachedHasher.arraysRoot([][]byte{}, 10, "testing")
+	_, err = NocachedHasher.arraysRoot([][]byte{}, 10, "testing")
 	assert.ErrorContains(t, "hash layer is a non power of 2", err)
 }
 
 func TestArraysTreeRoot_ZeroLength(t *testing.T) {
-	_, err := nocachedHasher.arraysRoot([][]byte{}, 0, "testing")
+	_, err := NocachedHasher.arraysRoot([][]byte{}, 0, "testing")
 	assert.ErrorContains(t, "zero leaves provided", err)
 }
