+++ testing/util/BUILD.bazel
@@ -26,6 +26,7 @@ go_library(
         "//beacon-chain/core/transition:go_default_library",
         "//beacon-chain/p2p/types:go_default_library",
         "//beacon-chain/state:go_default_library",
+        "//beacon-chain/state/stateutil:go_default_library",
         "//beacon-chain/state/v1:go_default_library",
         "//beacon-chain/state/v2:go_default_library",
         "//config/params:go_default_library",
